rename delimiter
